int/commands/commands.go: moves in the file.
