+++ internal/commands/commands.go
@@ -1,4 +1,4 @@
-package args
+package commands
 
 import (
 	"fmt"
