+++ main.go
@@ -4,11 +4,11 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/ufukty/bump/internal/args"
+	"github.com/ufukty/bump/internal/commands"
 )
 
 func main() {
-	if err := args.Dispatch(os.Args[1:]); err != nil {
+	if err := commands.Dispatch(os.Args[1:]); err != nil {
 		fmt.Println(err)
 		fmt.Println("try: bump help")
 		os.Exit(1)
